+++ src/de/omegazirkel/risingworld/landclaim/ui/PermissionOverlay.java
@@ -8,6 +8,7 @@ import net.risingworld.api.objects.Player;
 import net.risingworld.api.ui.style.Pivot;
 
 public class PermissionOverlay extends OZUIElement {
+    public static final String ATTRIBUTE_KEY = "landclaim-overlay";
 
     public PermissionOverlay(Area area, Player player, Callback<Player> onClose) {
         super();
@@ -17,11 +18,16 @@ public class PermissionOverlay extends OZUIElement {
         setBackgroundColor(0, 0, 0, 0.4f);
 
         AreaPermissionPanel panel = new AreaPermissionPanel(area, player, p -> {
-            p.removeUIElement(this);
-            CursorManager.hide(p);
-            onClose.onCall(p);
+            close(p, onClose);
         }, this);
         addChild(panel);
+    }
+
+    private void close(Player player, Callback<Player> onClose) {
+        player.removeUIElement(this);
+        player.deleteAttribute(ATTRIBUTE_KEY);
+        CursorManager.hide(player);
+        onClose.onCall(player);
     }
 
 }